requirements.txt
tkinter
pyautogui
Pillow
fpdf